gatekeeper: added handler and struct to config
Started working on the gatekeeper functionality.  Added struct for
gatekeeper options in config.go.  Also added the gatekeeper handler, and
registered it with the other bot handlers.  Finally, added fix for
privileged intents from the nickname-tracking branch.  Although, it was
slightly changed so that we are not overwriting a ton of values by
updating with a struct.  Instead we update value individually.