gatekeeper: handler, config, and new command
Added the update command, used to force the bot to update various
settings.  Currently, it is being used for the 'update gatekeeper'
command to update the welcome channel with new gatekeeper settings from
the config file.

Messed around with some things in cfg.go, but ultimately ended up
keeping everything the same for now.

Added a new handler but then it broke, so it is all commented out for
now.  Will work on fixing it.