gatekeeper: logic in place
Added the logic for the gatekeeper to the router.  I tried adding it in
gatekeeper.go again, as its own handler, but it is still is giving me a
ton of errors.  Adding the logic to the router, however, everything
works perfectly.

The gatekeeper portion of the router config is now being updated as well
from the config we read in when an 'update gk' command is used.