gatekeeper: basic functionality
Got the basic functionality of the gatekeeper working.  At the moment,
the logic which handles the gatekeeper is just baked into the router.
As such, there is no way to disable it.  Although, it can probably,
technically be disabled by just not having any gatekeeper values in your
config file.

Also updated the parameter names in the gatekeeper config to better
reflect what they represent (i.e. Channel is now ChannelID).